--- src/components/ForecastList/index.d.ts
@@ -1,5 +1,0 @@
-import * as React from "react";
-
-declare const ForecastList: React.ComponentType;
-
-export default ForecastList;
